requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
openai==1.12.0
gradio==4.19.2
redis==5.0.1
sqlalchemy==2.0.25
Pillow
python-multipart==0.0.9
pydantic==2.6.0
requests==2.31.0
python-dotenv==1.0.1
aiofiles==23.2.1